Split the rest of utils
